Moved Login components to own folder under Pages
